resources/dev/requirements.txt
pycryptodomex==3.22.0
pycryptodome==3.22.0
loguru==0.7.3
boto3==1.38.8
pyspark==3.5.5
findspark==2.0.1
mysql-connector-python==9.3.0
psycopg2-binary==2.9.10
botocore==1.38.8
jmespath==1.0.1
python-dateutil==2.9.0.post0
s3transfer==0.12.0
urllib3==2.4.0
faker==37.1.0
python-dotenv==1.1.0